backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
supabase==1.2.0
gotrue==1.3.0
httpx==0.24.1
anthropic==0.20.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
